(BSR)[PRO] refactor: move BoxFormLayout to ui kit
