+++ pro/src/components/UserEmailForm/UserEmailForm.tsx
@@ -4,8 +4,8 @@ import { api } from 'apiClient/api'
 import { isErrorAPIError } from 'apiClient/helpers'
 import { UserResetEmailBodyModel } from 'apiClient/v1'
 import { useCurrentUser } from 'commons/hooks/useCurrentUser'
-import { BoxFormLayout } from 'components/BoxFormLayout/BoxFormLayout'
 import { FormLayout } from 'components/FormLayout/FormLayout'
+import { BoxFormLayout } from 'ui-kit/BoxFormLayout/BoxFormLayout'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { PasswordInput } from 'ui-kit/form/PasswordInput/PasswordInput'
