+++ pro/src/components/UserIdentityForm/UserIdentityForm.tsx
@@ -6,8 +6,8 @@ import { api } from 'apiClient/api'
 import { isErrorAPIError } from 'apiClient/helpers'
 import { useCurrentUser } from 'commons/hooks/useCurrentUser'
 import { updateUser } from 'commons/store/user/reducer'
-import { BoxFormLayout } from 'components/BoxFormLayout/BoxFormLayout'
 import { FormLayout } from 'components/FormLayout/FormLayout'
+import { BoxFormLayout } from 'ui-kit/BoxFormLayout/BoxFormLayout'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { TextInput } from 'ui-kit/form/TextInput/TextInput'
