+++ pro/src/components/UserIdentityForm/__specs__/UserIdentity.spec.tsx
@@ -1,6 +1,5 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import React from 'react'
 
 import { api } from 'apiClient/api'
 import { sharedCurrentUserFactory } from 'commons/utils/factories/storeFactories'
