+++ pro/src/components/UserPasswordForm/UserPasswordForm.tsx
@@ -2,8 +2,8 @@ import { Form, FormikProvider, useFormik } from 'formik'
 
 import { api } from 'apiClient/api'
 import { isErrorAPIError } from 'apiClient/helpers'
-import { BoxFormLayout } from 'components/BoxFormLayout/BoxFormLayout'
 import { FormLayout } from 'components/FormLayout/FormLayout'
+import { BoxFormLayout } from 'ui-kit/BoxFormLayout/BoxFormLayout'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { PasswordInput } from 'ui-kit/form/PasswordInput/PasswordInput'
