+++ pro/src/components/UserPhoneForm/UserPhoneForm.tsx
@@ -8,8 +8,8 @@ import { UserPhoneBodyModel } from 'apiClient/v1'
 import { parseAndValidateFrenchPhoneNumber } from 'commons/core/shared/utils/parseAndValidateFrenchPhoneNumber'
 import { useCurrentUser } from 'commons/hooks/useCurrentUser'
 import { updateUser } from 'commons/store/user/reducer'
-import { BoxFormLayout } from 'components/BoxFormLayout/BoxFormLayout'
 import { FormLayout } from 'components/FormLayout/FormLayout'
+import { BoxFormLayout } from 'ui-kit/BoxFormLayout/BoxFormLayout'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { TextInput } from 'ui-kit/form/TextInput/TextInput'
