+++ pro/src/ui-kit/BoxFormLayout/components/BoxFormLayoutRequiredMessage.tsx
@@ -1,9 +1,8 @@
 import cn from 'classnames'
-import React from 'react'
 
 import { FormLayout } from 'components/FormLayout/FormLayout'
 
-import style from './BoxFormLayout.module.scss'
+import style from '../BoxFormLayout.module.scss'
 
 interface BoxFormLayoutRequiredMessageProps {
   className?: string
